+++ blockfrost-api/src/Blockfrost/Types/Shared/Slot.hs
@@ -9,7 +9,7 @@ import Servant.API (Capture, FromHttpApiData (..), ToHttpApiData (..))
 import Servant.Docs (DocCapture (..), ToCapture (..), ToSample (..), samples)
 
 newtype Slot = Slot Integer
-  deriving stock (Eq, Show, Generic)
+  deriving stock (Eq, Ord, Show, Generic)
   deriving newtype (Num, FromHttpApiData, ToHttpApiData, FromJSON, ToJSON)
 
 unSlot :: Slot -> Integer