+++ blockfrost-api/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Additions
   * `/blocks/${hash_or_number}/addresses` endpoint (Affected addresses in a block)
   * `blockTime` field to `AddressTransaction`
+  * `Ord` instances for `Address`, `AssetId`, `BlockHash`, `Epoch`, `Slot`, `TxHash`
 
 # Version [0.3.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.2.1.0...v0.3.0.0) (2022-02-07)